insert product into cart

+++ database/migrations/2024_05_07_144013_create_shopping_carts_table.php
@@ -1,4 +1,5 @@
 <?php
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -7,16 +8,22 @@ class CreateShoppingCartsTable extends Migration
 {
     public function up()
     {
-        Schema::create('shopping_carts', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('id_user');
-            $table->foreign('id_user')->references('id')->on('users')->onDelete('cascade');
-            $table->unsignedBigInteger('product_id');
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-            $table->integer('quantity')->default(0)->unsigned();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('shopping_carts')) {
+            Schema::create('shopping_carts', function (Blueprint $table) {
+                $table->id();
+                $table->bigInteger('id_user')->unsigned(); // Cambiado a bigInteger
+                $table->bigInteger('product_id')->unsigned(); // Cambiado a bigInteger
+                $table->integer('quantity')->default(0)->unsigned();
+                $table->timestamps();
+                $table->softDeletes();
+
+                $table->foreign('id_user')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+            });
+        }
     }
+
+
 
     public function down()
     {
